Add files via upload
The goal of this Bash script is to help administrators automate repetitive tasks that consume time. The idea behind this script is that you can easily add or remove options from the provided list, which contains 20 basic commands for administrators. If you have any suggestions, I would be honored if you could share your ideas so we can help others benefit from this.